+++ Example/StickyHeaderTabController/Tabs/TabData.swift
@@ -1,0 +1,73 @@
+//
+//  TabData.swift
+//  StickyHeaderTabController_Example
+//
+//  Created by Benjamin Chrobot on 9/26/17.
+//  Copyright © 2017 CocoaPods. All rights reserved.
+//
+
+import UIKit
+
+struct TabData {
+
+    static let states: [String] = [
+        "Alabama",
+        "Alaska",
+        "Arizona",
+        "Arkansas",
+        "California",
+        "Colorado",
+        "Connecticut",
+        "Delaware",
+        "Florida",
+        "Georgia",
+        "Hawaii",
+        "Idaho",
+        "Illinois Indiana",
+        "Iowa",
+        "Kansas",
+        "Kentucky",
+        "Louisiana",
+        "Maine",
+        "Maryland",
+        "Massachusetts",
+        "Michigan",
+        "Minnesota",
+        "Mississippi",
+        "Missouri",
+        "Montana",
+        "Nebraska",
+        "Nevada",
+        "New Hampshire",
+        "New Jersey",
+        "New Mexico",
+        "New York",
+        "North Carolina",
+        "North Dakota",
+        "Ohio",
+        "Oklahoma",
+        "Oregon",
+        "Pennsylvania",
+        "Rhode Island",
+        "South Carolina",
+        "South Dakota",
+        "Tennessee",
+        "Texas",
+        "Utah",
+        "Vermont",
+        "Virginia",
+        "Washington",
+        "West Virginia",
+        "Wisconsin",
+        "Wyoming",
+    ]
+
+    static let colors: [(name: String, color: UIColor)] = [
+        ("Red", .red),
+        ("Green", .green),
+        ("Blue", .blue),
+        ("Purple", .purple),
+        ("Magenta", .magenta),
+        ("Yellow", .yellow),
+    ]
+}