+++ StickyHeaderTabController/Classes/ViewControllers/StickyHeaderContentTabViewController.swift
@@ -1,0 +1,13 @@
+//
+//  StickyHeaderContentTabViewController.swift
+//  Treble
+//
+//  Created by Benjamin Chrobot on 9/20/17.
+//  Copyright © 2017 Treble Media, Inc. All rights reserved.
+//
+
+import UIKit
+
+open class StickyHeaderContentTabViewController: UIViewController {
+
+}